concretizer: rework operating system semantics for installed packages
The OS logic in the concretizer is still the way it was in the first version -- where defaults were
implemented in a fairly inflexible way using straight logic. Most of the other sections have been
reworked to leave these kinds of decisions to optimization -- so we do that for OS's as well.

As with targets, we optimize for target matches. We also try to optimize for OS matches between
nodes. Additionally, this commit adds the notion of "OS compatibility" where we allow for builds to
depend on binaries for certain other OS's. e.g, for macos, a bigsur build can depend on an already
installed (concrete) catalina build. One cool thing about this is that we can declare additional
compatible OS's later, e.g. CentOS and RHEL.